backend/requirements.txt
alembic==1.7.5
astroid==2.8.0
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.1
coverage==6.2
distlib==0.3.4
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==9.8.3
filelock==3.4.2
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-api-core==2.7.1
google-api-python-client==2.36.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.1.0
google-cloud-logging==2.7.0
google-cloud-pubsub==2.9.0
google-cloud-run==0.2.0
google-cloud-storage==1.42.3
google-cloud-tasks==2.8.1
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.56.0
greenlet==1.1.1
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.1
identify==2.4.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
libcst==0.3.21
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.2
oauth2client==4.1.3
openpyxl==3.0.9
packaging==21.3
pandas==1.3.3
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.16.0
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyflakes==2.4.0
pylint==2.11.1
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.5
pytest-env==0.6.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
uritemplate==3.0.1
urllib3==1.26.7
virtualenv==20.13.0
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.7.0
